clean up handling build results
